ADD TzaController

+++ src/main/java/com/competidor/springbootfundamentals/entity/Release.java
@@ -10,7 +10,7 @@ public class Release {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  private Integer id;
+  private Long id;
 
   private String releaseDate;
   private String description;
@@ -18,17 +18,17 @@ public class Release {
   public Release() {
   }
 
-  public Release(Integer id, String description, String releaseDate){
+  public Release(Long id, String description, String releaseDate){
     this.id = id;
     this.releaseDate = releaseDate;
     this.description = description;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
