+++ src/main/java/com/competidor/springbootfundamentals/entity/Ticket.java
@@ -3,12 +3,13 @@ package com.competidor.springbootfundamentals.entity;
 
 import javax.persistence.*;
 
+
 @Entity
 public class Ticket {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  private Integer id;
+  private Long id;
 
   private String title;
   private String description;
@@ -43,11 +44,11 @@ public class Ticket {
     this.status = status;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
